add crud company

+++ routes/web.php
@@ -1,12 +1,13 @@
 <?php
 
 // use App\Http\admin\Controllers\OfferController as ControllersOfferController;
-
-use App\Http\Controllers\CityController;
-use App\Http\Controllers\OfferController;
-use App\Http\Controllers\ProfileController;
-use App\Http\Controllers\UserController;
 use Illuminate\Support\Facades\Route;
+
+use App\Http\Controllers\Ceo\CompanyController;
+use App\Http\Controllers\ProfileController;
+use App\Http\Controllers\Admin\CityController;
+use App\Http\Controllers\Admin\UserController;
+use App\Http\Controllers\Admin\OfferController;
 
 /*
 |--------------------------------------------------------------------------
@@ -22,7 +23,7 @@ use Illuminate\Support\Facades\Route;
 Route::get('/', function () {
     return view('welcome');
 });
-Route::resource('users',UserController::class);
+Route::namespace('Admin')->resource('users',UserController::class);
 
 Route::get('/dashboard', function () {
     return view('dashboard');
@@ -39,23 +40,35 @@ Route::get('/master',function(){
 });
 
 // Route users:
-Route::get('/dashboard', [UserController::class , 'allusers'])->name('dashboard');
-Route::delete('/admin/{user}', [UserController::class, 'destroy'])->name('users.destroy');
-Route::get('/admin/{user}/edit', [UserController::class, 'edit'])->name('users.edit');
-Route::put('/admin/{user}', [UserController::class, 'update'])->name('users.update');
-Route::put('restore/{user}', [UserController::class, 'restore'])->name('users.restore');
+Route::namespace('Admin')->get('/dashboard', [UserController::class , 'allusers'])->name('dashboard');
+Route::namespace('Admin')->delete('/admin/{user}', [UserController::class, 'destroy'])->name('users.destroy');
+Route::namespace('Admin')->get('/admin/{user}/edit', [UserController::class, 'edit'])->name('users.edit');
+Route::namespace('Admin')->put('/admin/{user}', [UserController::class, 'update'])->name('users.update');
+Route::namespace('Admin')->put('restore/{user}', [UserController::class, 'restore'])->name('users.restore');
 
 
 // Route::get('/ajoute',function(){
 //     return view('admin.AjouteCaractaire');
 // });
-
-Route::resource('cities',CityController::class);
-Route::get('/offer',[OfferController::class,'index'])->name('offer.index');
-Route::get('/offer/create',[OfferController::class,'create'])->name('offres.create');
-Route::post('/offer/store',[OfferController::class,'store'])->name('offres.store');
+Route::get('/home',function(){
+    return view('home');
+});
 
 
+
+// Route cites
+Route::namespace('Admin')->resource('cities',CityController::class);
+
+
+// Route Jobs:
+Route::namespace('Admin')->get('/offer',[OfferController::class,'index'])->name('offer.index');
+Route::namespace('Admin')->get('/offer/create',[OfferController::class,'create'])->name('offres.create');
+Route::namespace('Admin')->post('/offer/store',[OfferController::class,'store'])->name('offres.store');
+
+
+
+//route ceo
+    Route::namespace('Ceo')->resource('company', CompanyController::class);
 
 
 require __DIR__.'/auth.php';
